drivers: rcar_mmc: enable support of SDR104 timing mode
Add max freq correction to host properties f_max. SD/MMC clock for
Gen3/Gen4 R-car boards can't be equal to 208 MHz, but we can run SDR104
on lower frequencies according to SD card standard, and it is needed to
pass check of frequency in SDMMC subsystem in order to use SDR104 mode.

Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>
Acked-by: Dmytro Firsov <dmytro_firsov@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>